refactor: refactor project, add base documentation
