+++ src/components/Canvas/utils/imageUtils.js
