+++ src/components/Canvas/utils/textUtils.js
